requirements.txt
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
ase==3.25.0
async-timeout==5.0.1
attrs==25.3.0
bitsandbytes==0.45.5
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cut-cross-entropy==25.1.1
cycler==0.12.1
datasets==3.6.0
diffusers==0.33.1
dill==0.3.8
docstring-parser==0.16
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.0
gguf==0.16.3
hf-transfer==0.1.9
hf-xet==1.1.2
huggingface-hub==0.32.0
idna==3.10
importlib-metadata==8.7.0
jarvis-tools==2024.10.30
jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
msgspec==0.19.0
multidict==6.4.4
multiprocess==0.70.16
mypy-extensions==1.1.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
peft==0.15.2
pillow==11.2.1
platformdirs==4.3.8
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
pyarrow==20.0.0
pydantic==2.11.5
pydantic-core==2.33.2
pydantic-settings==2.9.1
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentencepiece==0.2.0
setuptools==80.8.0
shtab==1.7.2
six==1.17.0
spglib==2.6.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
toolz==1.0.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
triton==3.3.0
trl==0.15.2
typeguard==4.4.2
typing-extensions==4.13.2
typing-inspection==0.4.1
tyro==0.9.21
tzdata==2025.2
urllib3==2.4.0
uv==0.7.8
wheel==0.45.1
xformers==0.0.30
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
